pyproject.toml
[project]
name = "heretix"
version = "0.1.0"
description = "Heretix RPL harness (clean package)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "openai>=1.40",
    "requests>=2.32",
    "tldextract>=5.1",
    "beautifulsoup4>=4.12",
    "python-dateutil>=2.9",
    "filelock>=3.15",
    "pydantic>=2.6",
    "pydantic-settings>=2.5",
    "email-validator>=2.2",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "jsonschema>=4.22",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "alembic>=1.13",
    "fastapi>=0.116",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "stripe>=11.0",
    "duckdb>=1.4.1",
    "google-cloud-storage>=3.4.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0,<7.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0,<7.0",
    "pre-commit>=3.0",
]

[project.scripts]
heretix = "heretix.cli:app"

[tool.setuptools.packages.find]
include = ["heretix*", "heretix_wel*", "heretix_api*"]
exclude = ["legacy*", "heretix_rpl*", "heretix_promptstudio*", "runs*", "tests*"]

[tool.uv]
package = true
